thread the state down through the <App/>
We thread the state of the app from the top root component <App/>
through any intermediate components (here <TodoList/>) and finally down
into the leaf components, in this case the <Todo/> which represents a
single item.

At this point we will throw an error in the browser because the state
object does not exist.